Access control changes.

+++ protected/controllers/TmanageController.php
@@ -20,26 +20,15 @@ class TmanageController extends Controller
 	}
 
 	/**
-	 * Specifies the access control rules.
-	 * This method is used by the 'accessControl' filter.
-	 * @return array access control rules
-	 */
-	public function accessRules()
-	{
-		return array(
-			array('allow',  // allow all users to perform 'index' and 'view' actions
-				'actions'=>array('index','view','create','update','admin','delete'),
-				'users'=>array('*'),
-			),
-		);
-	}
-
-	/**
 	 * Displays a particular model.
 	 * @param integer $id the ID of the model to be displayed
 	 */
 	public function actionView($id)
 	{
+        if (Yii::app()->user->getId() != 49) {
+            throw new CHttpException(403, 'У вас немає права редагування цього документа.');
+        }
+
 		$this->render('view',array(
 			'model'=>$this->loadModel($id),
 		));
@@ -51,6 +40,10 @@ class TmanageController extends Controller
 	 */
 	public function actionCreate()
 	{
+        if (Yii::app()->user->getId() != 49) {
+            throw new CHttpException(403, 'У вас немає права редагування цього документа.');
+        }
+
 		$model=new Teacher;
 
 		// Uncomment the following line if AJAX validation is needed
@@ -75,6 +68,9 @@ class TmanageController extends Controller
 	 */
 	public function actionUpdate($id)
 	{
+        if (Yii::app()->user->getId() != 49) {
+            throw new CHttpException(403, 'У вас немає права редагування цього документа.');
+        }
 		$model=$this->loadModel($id);
 
 		// Uncomment the following line if AJAX validation is needed
@@ -99,6 +95,10 @@ class TmanageController extends Controller
 	 */
 	public function actionDelete($id)
 	{
+        if (Yii::app()->user->getId() != 49) {
+            throw new CHttpException(403, 'У вас немає права редагування цього документа.');
+        }
+
 		$this->loadModel($id)->delete();
 
 		// if AJAX request (triggered by deletion via admin grid view), we should not redirect the browser
@@ -111,6 +111,10 @@ class TmanageController extends Controller
 	 */
 	public function actionIndex()
 	{
+        if (Yii::app()->user->getId() != 49) {
+            throw new CHttpException(403, 'У вас немає права редагування цього документа.');
+        }
+
 		$dataProvider=new CActiveDataProvider('Teacher');
 		$this->render('index',array(
 			'dataProvider'=>$dataProvider,
@@ -122,6 +126,10 @@ class TmanageController extends Controller
 	 */
 	public function actionAdmin()
 	{
+        if (Yii::app()->user->getId() != 49) {
+            throw new CHttpException(403, 'У вас немає права редагування цього документа.');
+        }
+
 		$model=new Teacher('search');
 		$model->unsetAttributes();  // clear any default values
 		if(isset($_GET['Teacher']))
